# apidoc/dashboard-openapi.yaml
openapi: 3.0.0
info:
  title: Spotlight Dashboard REST API
  version: 1.0.0-20190117
servers:
- url: http://localhost:9000/dashapi/v1
...
security:
- bearer: []
paths:
...
  /users/:
    get:
      summary: Returns list of users
      tags:
      - Users
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Creates and returns a new user
      tags:
      - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "User 1"
                sso:
                  type: integer
                  example: 999999999
                roles:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
      responses:
        201:
          description: Successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userId}:
    get:
      summary: Returns the user with userId
      tags:
      - Users
      parameters:
      - name: userId
        in: path
        description: User id
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      summary: Updates the user name for the user with userId.
      tags:
      - Users
      parameters:
      - name: userId
        in: path
        description: User id
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'User 1'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Deletes the user with userId.
      description: Deletes the user with userId. It won't delete the user if the user is added to at least one process.
      tags:
      - Users
      parameters:
      - name: userId
        in: path
        description: User id
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /processes/{processId}/users:
    get:
      summary: Returns all the users added to the process with processId.
      tags:
      - Processes
      parameters:
      - name: processId
        in: path
        description: Process id
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProcessUser'
    post:
      summary: Add a new user to the process with processId.
      tags:
      - Processes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  example: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessUser'
  /processes/{processId}/users/{userId}:
    delete:
      summary: Removes a user from the process with processId.
      tags:
      - Processes
      parameters:
      - name: processId
        in: path
        description: Process id
        required: true
        schema:
          type: integer
      - name: userId
        in: path
        description: User id
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessUser'
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  schemas:
...
    ProcessUser:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user:
          $ref: '#/components/schemas/User'
        processId:
          type: integer
          example: 1
...
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: "User 1"
        sso:
          type: integer
          example: 999999999
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
...